mac: Permit Chrome to run from the data volume on macOS 11 and 12
A sandbox bug on macOS 11 and 12 causes the sandbox to see any paths
within the data volume outside of system firm link locations (listed in
/usr/share/firmlinks) as though they were not on the data volume,
causing it to deny access to user paths containing the data volume mount
point as a prefix. This is filed as
https://openradar.appspot.com/FB9738355.

When Chrome's bundle path appears in the data volume, this causes the
sandbox to deny access to the bundle.

This is not a problem in normal use, as typical bundle paths, while on
the data volume, will be in system firm link locations such as
/Applications or /Users. As a workaround for other cases where the
bundle may be present on the data volume but not in a system firm link
location, configure the sandbox with an alternate bundle path so that it
permits access to the bundle.

Bug: 1266490
Change-Id: Idad66bf16af031dbb52dab7c61a96011748afabc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3373365
Auto-Submit: Mark Mentovai <mark@chromium.org>
Reviewed-by: Robert Sesek <rsesek@chromium.org>
Commit-Queue: Robert Sesek <rsesek@chromium.org>
Cr-Commit-Position: refs/heads/main@{#957208}
NOKEYCHECK=True
GitOrigin-RevId: cc5618e68a210cf50053406e1a46d2af3d09dab6