Prevent sandbox::TestRunner from reusing Policy
In Chrome we only use each Policy for a single process, while
in tests we often apply the same policy to multiple processes.

The critical part of this change is the CHECK in controller.cc,
everything else is introducing fresh |TestRunner|s for each call
to RunTest(). Where the runner setup was simple we introduce
multiple TestRunner within each TEST() function. Where the setup
is more complex we provide a helper returning a
std::unique_ptr<TestRunner>() and call it multiple times to reset
the runner.

Bug: 1270309,1291884
Change-Id: I353051c59e0fb43187961f435530795abd81fc08
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3426302
Reviewed-by: Will Harris <wfh@chromium.org>
Commit-Queue: Alex Gough <ajgo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#967561}
NOKEYCHECK=True
GitOrigin-RevId: 137d2680a17eb69b63130b0015407ae105cb3cf2