Add more launch failure codes for sandboxed process start.
This CL adds a few more checks on a sandboxed target process.
In particular, it will verify that the process is launched
by the main exe (moves this check from a DCHECK previously).

It will also now check to a reasonable degree that the target
process is the same process image as the broker process, by
verifying that it's loaded at the same address, and also
verifying that sentinel values around the variable transfer
locations are valid.

This should allow determination of whether the flood of
SBOX_ERROR_CANNOT_WRITE_VARIABLE_VALUE errors are caused
by any of these conditions, or something else.

This CL also fixes a test-only function to have a conformant
name.

BUG=1314491

Change-Id: I98696d5a8eb88254d01c5ced121250b983b620e6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3583026
Reviewed-by: Alex Gough <ajgo@chromium.org>
Commit-Queue: Will Harris <wfh@chromium.org>
Cr-Commit-Position: refs/heads/main@{#992636}
NOKEYCHECK=True
GitOrigin-RevId: e82d487220eefb90aa34310fb09674da6d7ffadc