Annotate NetworkContext with RequireContext= attribute
Adds sandbox.mojom.Context - an enum to represent broad privilege
levels or contexts within Chromium. These levels are not the same
as sandbox types but instead describe the capabilities that can
be allowed in a given process or service. A tightly sandboxed
utility might have access to a powerful capability (as here,
the NetworkContext in the Cast mirroring service), but all other
tightly sandboxed services should not.

This enum is then used to annotate the NetworkContext - a powerful
remote to an object in the network service that allows wide access
to a user's cookies and connections. The NetworkContext is marked
as kPrivilegedUtility as it is required (at present) by the cast
mirroring service as well as the browser process. (It is quite
possible that it might not need this full capability.)

Various methods that pass this interface are then marked with
the context in which they are expected to be called in Chromium.
For now the only check that these contexts are honored is at
mojom generation time but future CLs may add CHECKs or DCHECKs
to the remotes or proxies to add an extra layer of enforcement.

Bug: 931315
Change-Id: I73aed8cfc1a8516922515249f62785e9e3c79b2d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3589258
Reviewed-by: Robert Sesek <rsesek@chromium.org>
Commit-Queue: Alex Gough <ajgo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#994740}
NOKEYCHECK=True
GitOrigin-RevId: 4a96040a1ede099d3d9706a97c3221d4e8cbc044