[fuchsia] Run sub-process FilteredServiceDirectory on its own thread
This moves sub-processes FilteredServiceDirectory on their own
separate IO threads. This fixes a number on browser tests that attempt
to run synchronous calls to other processes (such as the network
process) as part of their setup flow.

Currently, the FilteredServiceDirectories for sub-processes are
handled in the main IO thread for the browser process. As a result,
when sub-processes attempt to use an external service, they hang
forever since the IO thread is blocked in the browser process.

This fixes the issue by handling the FilteredServiceDirectories for
sub-processes on separate threads, freeing up the main IO thread in
the browser process.

Bug: 1301356, 1303038
Change-Id: I087c38dfc66d5842b19558c54c8cdacc83b095e4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3506648
Reviewed-by: Wez <wez@chromium.org>
Auto-Submit: Fabrice de Gans <fdegans@chromium.org>
Reviewed-by: Eric Seckler <eseckler@chromium.org>
Commit-Queue: Eric Seckler <eseckler@chromium.org>
Cr-Commit-Position: refs/heads/main@{#994853}
NOKEYCHECK=True
GitOrigin-RevId: e08d4ea3103d7805adb90120fc05ee8b5f832f99