Bump minimum version supported for App Container to RS5.
This is because it is not possible to set process mitigations
on an App Container process until RS5.

Since the number of users in the range RS2 <-> RS4 is small
it's not worth trying to fix or support this configuration.

Better to bump the minimum version to RS5 and all the APIs
are fully supported.

BUG=1318654

Change-Id: I22d958232e4bc1a489fb4d6e45cc6e182b42aed7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3600767
Reviewed-by: Alex Gough <ajgo@chromium.org>
Commit-Queue: Will Harris <wfh@chromium.org>
Cr-Commit-Position: refs/heads/main@{#995046}
NOKEYCHECK=True
GitOrigin-RevId: fa710da7509419524f1f96fc1cae003ba223962c