Skeleton for shared backing data for sandbox policies
This CL adds a new //sandbox/win interface TargetConfig (with no
useful methods in this CL) which will gradually replace the methods
offered by TargetPolicy. ConfigBase implements this new interface.

BrokerServices gains a new overloaded method CreatePolicy(tag)
which creates a PolicyBase with a backing TargetConfig from
a cache keyed by the opaque string |tag|. TargetConfig knows if it has
been configured or not, and sandbox delegates can query this to
determine if its methods should be called (the first time a policy
corresponding to a tag is seen) or not (subsequent times).

Tests are added to validate that policies are created and fixed as
expected. Nothing outside of //sandbox/win makes use of this new
interface yet.

Bug: 549319
Change-Id: I8ce788052e92f01ad45f3a376a1d2d7bf1524e39
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3556095
Commit-Queue: Alex Gough <ajgo@chromium.org>
Reviewed-by: Will Harris <wfh@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1027414}
NOKEYCHECK=True
GitOrigin-RevId: dec8803c437a80ef23cf33a91cc6107a4b41846d