Share LowLevelPolicy between sandboxed processes
This moves LowLevelPolicy (and PolicyMaker) to shared backing storage
allowing it to be shared between sandbox policies that opt-in to
sharing.

AddRule is moved to the TargetConfig interface as it now only needs
to be called the first time a policy with a given tag is configured.

Delegates/Embedders can opt out of sharing by supplying an
empty sandbox tag so that the backing storage is not used.

AddGenericDllEvictionPolicy() is hoisted out of AddGenericPolicy()
as it is not yet part of the fixed part of the policy rules.

This should reduce the amount of work needed when setting up
individual processes, and will share policy rules memory. This
is ok as the rules memory is location agnostic in both the broker
and child (target), and is not written to once it is compiled for
the first launch.

Bug: 549319
Change-Id: I249fef69296adb2ed289f2a00b122763d613fda8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3558444
Reviewed-by: Bo Liu <boliu@chromium.org>
Reviewed-by: danakj <danakj@chromium.org>
Commit-Queue: Alex Gough <ajgo@chromium.org>
Reviewed-by: Daniel Rubery <drubery@chromium.org>
Reviewed-by: Will Harris <wfh@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1027587}
NOKEYCHECK=True
GitOrigin-RevId: 6cde7d618878230861163e0d68ba652549ba6a45