lacros: Propagate --gpu-sandbox-start-early from ash-chrome.
This CL propagates the --gpu-sandbox-start-early switch from ash-chrome
to lacros-chrome: on ChromeOS, some boards need to start the GPU process
sandbox early due to the presence of multiple threads at the regular
starting point.

Propagating this switch allows us to make GPU sandbox failures fatal, so
we do that as well in this CL.

Version skew considerations: if lacros-chrome runs with an ash-chrome
that's too old to propagate the switch, we start the sandbox early -- we
consider this to be better than not starting the GPU process or starting
it without a sandbox.

Bug: 1339079
Test: inspect chrome://gpu for lacros on volteer to ensure sandbox is on
Test: inspect chrome://gpu for lacros on atlas to ensure sandbox is on
Change-Id: I2d45f2aed56c8a3548f4224244a39410eafd838d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3817593
Reviewed-by: Kenneth Russell <kbr@chromium.org>
Reviewed-by: Ken Rockot <rockot@google.com>
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Commit-Queue: Andres Calderon Jaramillo <andrescj@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1036909}
NOKEYCHECK=True
GitOrigin-RevId: eaf54bd8bcd0212913a93b18945def5fdb9d3435