Move SetAlternateDesktop to sandbox config
NOTE: CHANGES BEHAVIOR of SetAlternateDesktop / CreateAlternateDesktop.

This moves TargetPolicy::SetAlternateDesktop(bool) to the
TargetConfig::SetDesktop(enum Desktop) interface, but leaves the
remaining functions that manipulate the backing desktop and
winstation handles on TargetPolicy. In a future change they may move
to Broker Services as they are not strictly a policy configuration
but manipulate global sandbox state.

After this change alternate desktops are NO LONGER created
during calls to SetAlternateDesktop but must be pre-created
before any calls to SetDesktop()/GetDesktopName() or (transitively)
BrokerServices::SpawnTarget.

In Chrome this already happens[0] as we call:

```
 broker->CreatePolicy()->CreateAlternateDesktop(...);
```

in Sandbox::Initialize().

Various tests did not pre-create the desktop, and we
do so now.

[0] https://source.chromium.org/chromium/chromium/src/+/main:sandbox/policy/sandbox.cc;drc=d1565c32ac36d208c2c7d71ad2fe21b26c2812c5;l=60

Summary:
 bool alternate_winstation => enum Desktop.

 TargetPolicy::SetAlternateDesktop => TargetConfig::SetDesktop
  -> no longer creates desktops
 TargetPolicy::DestroyAlternateDesktops => DestroyDesktops
  -> destroys all backing static desktops
 TargetPolicy::GetAlternateDesktop => GetDesktopName
  -> shorter name, no functional change.

Bug: 549319
Change-Id: Ic230705cfd0c78ecd6fa8b79b99de3db737725c1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3843162
Commit-Queue: Alex Gough <ajgo@chromium.org>
Reviewed-by: Will Harris <wfh@chromium.org>
Reviewed-by: Daniel Rubery <drubery@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1044797}
NOKEYCHECK=True
GitOrigin-RevId: fac368c7bb91dac1fcddb6f4415ff10b4074d478