Enable ChromeScreenAI on Windows.
ChromeScreenAI is enabled on Windows.

Sandbox is set to allow read only access to installed component folder,
this will be removed before launch and replaced with opening required
files in the browser and sending their handles to the ScreenAI binary.

The component binary path is now stored in InstallState and passed to
the service after service is initialized.

Bug: 1278249
Change-Id: Ia86e3f6309e55195040ab1927d54ca4f9018ed6f
Ax-Relnotes: n/a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3876149
Reviewed-by: Sorin Jianu <sorin@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: Alex Gough <ajgo@chromium.org>
Reviewed-by: David Tseng <dtseng@chromium.org>
Commit-Queue: Ramin Halavati <rhalavati@chromium.org>
Reviewed-by: Matthew Denton <mpdenton@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1060869}
NOKEYCHECK=True
GitOrigin-RevId: fd712d8e309111a0660b74ee6dcd42efa383a40a